+++ cwltool/process.py
@@ -209,7 +209,7 @@ def stageFiles(pm, stageFunc, ignoreWritable=False):
             if p.type == "File":
                 shutil.copy(p.resolved,p.target)
             elif p.type == "Directory":
-                if os.path.exists(p.target):
+                if os.path.exists(p.target) and os.path.isdir(p.target):
                     shutil.rmtree(p.target)
                 shutil.copytree(p.resolved, p.target)
         elif p.type == "Directory" and not os.path.exists(p.target) and p.resolved.startswith("_:"):
